requirements.txt
numpy==1.21.5
opencv-python==4.5.4.60
pandas==1.3.5
Pillow==8.4.0
plyfile==0.7.4
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
torch==1.10.1
typing_extensions==4.0.1
